Place the new test logic at WorkspaceExecuteCommandHandlerTest
