+++ org.eclipse.jdt.ls.tests/src/org/eclipse/jdt/ls/core/internal/handlers/WorkspaceExecuteCommandHandlerTest.java
@@ -12,9 +12,14 @@
 package org.eclipse.jdt.ls.core.internal.handlers;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 
 import java.util.Arrays;
+import java.util.Set;
 
+import org.eclipse.jdt.ls.core.internal.JavaLanguageServerPlugin;
 import org.eclipse.jdt.ls.core.internal.managers.AbstractProjectsManagerBasedTest;
 import org.eclipse.lsp4j.ExecuteCommandParams;
 import org.eclipse.lsp4j.jsonrpc.ResponseErrorException;
@@ -23,6 +28,8 @@ import org.hamcrest.Matcher;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
+import org.osgi.framework.Bundle;
+import org.osgi.framework.BundleContext;
 
 public class WorkspaceExecuteCommandHandlerTest extends AbstractProjectsManagerBasedTest {
 
@@ -105,5 +112,42 @@ public class WorkspaceExecuteCommandHandlerTest extends AbstractProjectsManagerB
 		WorkspaceExecuteCommandHandler handler = WorkspaceExecuteCommandHandler.getInstance();
 		ExecuteCommandParams params = null;
 		handler.executeCommand(params, monitor);
+	}
+
+	@Test
+	public void testRegistryEventListener() throws Exception {
+		loadBundles(Arrays.asList(getBundle("testresources", "jdt.ls.extension-0.0.1.jar")));
+		String bundleLocation = getBundleLocation(getBundle("testresources", "jdt.ls.extension-0.0.1.jar"), true);
+
+		BundleContext context = JavaLanguageServerPlugin.getBundleContext();
+		Bundle installedBundle = context.getBundle(bundleLocation);
+		try {
+			assertNotNull(installedBundle);
+
+			assertTrue(installedBundle.getState() == Bundle.STARTING || installedBundle.getState() == Bundle.ACTIVE);
+			installedBundle.loadClass("jdt.ls.extension.Activator");
+			assertEquals(installedBundle.getState(), Bundle.ACTIVE);
+
+			Set<String> extensionCommands = WorkspaceExecuteCommandHandler.getInstance().getAllCommands();
+			assertTrue(extensionCommands.contains("jdt.ls.extension.command1"));
+			assertTrue(extensionCommands.contains("jdt.ls.extension.command2"));
+
+			loadBundles(Arrays.asList(getBundle("testresources", "jdt.ls.extension-0.0.2.jar")));
+			bundleLocation = getBundleLocation(getBundle("testresources", "jdt.ls.extension-0.0.2.jar"), true);
+
+			installedBundle = context.getBundle(bundleLocation);
+			assertNotNull(installedBundle);
+			assertTrue(installedBundle.getState() == Bundle.STARTING || installedBundle.getState() == Bundle.ACTIVE);
+			installedBundle.loadClass("jdt.ls.extension.Activator");
+			assertEquals(installedBundle.getState(), Bundle.ACTIVE);
+
+			extensionCommands = WorkspaceExecuteCommandHandler.getInstance().getAllCommands();
+			assertTrue(extensionCommands.contains("jdt.ls.extension.command2"));
+			assertTrue(extensionCommands.contains("jdt.ls.extension.command3"));
+			assertFalse(extensionCommands.contains("jdt.ls.extension.command1"));
+		} finally {
+			// Uninstall the bundle to clean up the testing bundle context.
+			installedBundle.uninstall();
+		}
 	}
 }
